requirement.txt
appdirs==1.4.4
asgiref==3.3.1
black==20.8b1
certifi==2020.11.8
click==7.1.2
crispy-forms-gds==0.2.1
Django==3.1.4
django-ckeditor==6.0.0
django-ckeditor-widget==0.1
django-countries==6.1.3
django-crispy-forms==1.10.0
django-js-asset==1.2.2
django-phonenumber-field==5.0.0
mypy-extensions==0.4.3
pathspec==0.8.1
phonenumbers==8.12.13
Pillow==8.0.1
pytz==2020.4
regex==2020.11.13
repackage==0.5
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
wincertstore==0.2
